+++ src/main/java/com/accbdd/complicated_bees/block/AbstractMellariumBlock.java
@@ -118,13 +118,4 @@ public abstract class AbstractMellariumBlock extends BaseEntityBlock {
         super.createBlockStateDefinition(pBuilder);
         pBuilder.add(EsotericRegistration.ASSEMBLED);
     }
-
-    public enum MellariumBlockType {
-        BASE,
-        CONTROLLER,
-        FAN,
-        HEATER,
-        HYDRO,
-        FRAME
-    }
 }