+++ src/main/java/com/accbdd/complicated_bees/block/AbstractMellariumPoweredBlock.java
@@ -1,0 +1,21 @@
+package com.accbdd.complicated_bees.block;
+
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import org.jetbrains.annotations.Nullable;
+
+public abstract class AbstractMellariumPoweredBlock extends MellariumBlock {
+    @Override
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> pBuilder) {
+        super.createBlockStateDefinition(pBuilder);
+        pBuilder.add(BlockStateProperties.POWERED);
+    }
+
+    @Override
+    public @Nullable BlockState getStateForPlacement(BlockPlaceContext pContext) {
+        return super.getStateForPlacement(pContext).setValue(BlockStateProperties.POWERED, false);
+    }
+}