+++ src/main/java/com/accbdd/complicated_bees/datagen/BlockTagGenerator.java
@@ -38,7 +38,8 @@ public class BlockTagGenerator extends BlockTagsProvider {
                 BlocksRegistration.MELLARIUM_RAIN_SHIELD.get(),
                 BlocksRegistration.MELLARIUM_MUTATOR.get(),
                 BlocksRegistration.MELLARIUM_HYDROREGULATOR.get(),
-                BlocksRegistration.MELLARIUM_ENERGY_ACCEPTOR.get()
+                BlocksRegistration.MELLARIUM_ENERGY_ACCEPTOR.get(),
+                BlocksRegistration.MELLARIUM_SKYBOX.get()
         );
         tag(BlockTags.MINEABLE_WITH_AXE).add(
                 BlocksRegistration.APIARY.get(),