+++ src/main/java/com/accbdd/complicated_bees/datagen/ItemModelGenerator.java
@@ -37,6 +37,7 @@ public class ItemModelGenerator extends ItemModelProvider {
         withExistingParent(BlocksRegistration.MELLARIUM_MUTATOR.getId().getPath(), modLoc("block/mellarium_mutator"));
         withExistingParent(BlocksRegistration.MELLARIUM_HYDROREGULATOR.getId().getPath(), modLoc("block/mellarium_hydroregulator"));
         withExistingParent(BlocksRegistration.MELLARIUM_ENERGY_ACCEPTOR.getId().getPath(), modLoc("block/mellarium_energy_acceptor"));
+        withExistingParent(BlocksRegistration.MELLARIUM_SKYBOX.getId().getPath(), modLoc("block/mellarium_skybox"));
         microscopeModel();
         withExistingParent(BlocksRegistration.WAX_BLOCK.getId().getPath(), modLoc("block/wax_block"));
         withExistingParent(BlocksRegistration.WAX_BLOCK_STAIRS.getId().getPath(), modLoc("block/wax_block_stairs"));