+++ src/main/java/com/accbdd/complicated_bees/datagen/loot/BlockLootTables.java
@@ -45,6 +45,7 @@ public class BlockLootTables extends BlockLootSubProvider {
         dropSelf(BlocksRegistration.MELLARIUM_MUTATOR.get());
         dropSelf(BlocksRegistration.MELLARIUM_HYDROREGULATOR.get());
         dropSelf(BlocksRegistration.MELLARIUM_ENERGY_ACCEPTOR.get());
+        dropSelf(BlocksRegistration.MELLARIUM_SKYBOX.get());
         this.add(BlocksRegistration.BEE_NEST.get(), nestLootTable(BlocksRegistration.BEE_NEST.get()));
         dropSelf(BlocksRegistration.WAX_BLOCK.get());
         dropSelf(BlocksRegistration.WAX_BLOCK_STAIRS.get());