+++ src/main/java/com/accbdd/complicated_bees/registry/BlocksRegistration.java
@@ -44,6 +44,7 @@ public class BlocksRegistration {
     public static final RegistryObject<MellariumBlock> MELLARIUM_MUTATOR = BLOCKS.register("mellarium_mutator", MellariumMutatorBlock::new);
     public static final RegistryObject<MellariumBlock> MELLARIUM_HYDROREGULATOR = BLOCKS.register("mellarium_hydroregulator", MellariumHydroregulatorBlock::new);
     public static final RegistryObject<MellariumBlock> MELLARIUM_ENERGY_ACCEPTOR = BLOCKS.register("mellarium_energy_acceptor", MellariumEnergyAcceptorBlock::new);
+    public static final RegistryObject<MellariumBlock> MELLARIUM_SKYBOX = BLOCKS.register("mellarium_skybox", MellariumSkyboxBlock::new);
     public static final RegistryObject<MellariumBlock> MELLARIUM_CONTROLLER = BLOCKS.register("mellarium_controller", () -> new MellariumBlock(MellariumBlock.MellariumBlockType.CONTROLLER));
 
     public static final RegistryObject<Block> APID_LIBRARY = BLOCKS.register("apid_library", ApidLibraryBlock::new);